Add files via upload
Training documentation/Job Aid for PGE Division individuals working with Website to verify and update Project polygon areas.  There is a copy of this that is live and subject to updates as needed.  This is a snapshot captured December 3, 2021.  This is associated with and in support of the two training sessions where one of these sessions was captured on video and an online training session was created.